refactor(skills): extract capability layer from runtime
