+++ src/voidcode/skills/__init__.py
@@ -1,0 +1,27 @@
+from .discovery import (
+    DEFAULT_SKILL_SEARCH_PATHS,
+    SKILL_ENTRY_FILE_NAME,
+    LocalSkillMetadataLoader,
+    resolve_workspace_relative_path,
+)
+from .manifest import (
+    FRONTMATTER_DELIMITER,
+    SUPPORTED_FRONTMATTER_KEYS,
+    parse_skill_body,
+    parse_skill_frontmatter,
+)
+from .models import SkillMetadata
+from .registry import SkillRegistry
+
+__all__ = [
+    "DEFAULT_SKILL_SEARCH_PATHS",
+    "FRONTMATTER_DELIMITER",
+    "LocalSkillMetadataLoader",
+    "SKILL_ENTRY_FILE_NAME",
+    "SUPPORTED_FRONTMATTER_KEYS",
+    "SkillMetadata",
+    "SkillRegistry",
+    "parse_skill_body",
+    "parse_skill_frontmatter",
+    "resolve_workspace_relative_path",
+]
